refactor(api): move streamSSEUtf8 out of services/llm/ [CR-SSE-LOCATION-15]
`streamSSEUtf8` is an HTTP-layer wrapper around Hono's `streamSSE` —
it has nothing to do with LLMs. Living under `services/llm/` was a
historical accident from when the only SSE consumer was the LLM
streaming route. Now both `routes/interview.ts` and `routes/sessions.ts`
import it (and future per-message-tool routes will too), so the
wrapper belongs in a transport-level location.

Moved to `services/streaming/sse-utf8.ts`. No behavior change — only
the import path changes in the two call sites. Helper docstring (the
mojibake / Latin-1 fallback explanation from BUG-881) is preserved
verbatim.
